refactor azblob
move generated code to internal package
simplified codegen to contain only request creation/response handling
internal types are exposed via type aliasing in the public package
updated page iterator per latest discussion
